Added:
-RenderGraph, a helper class for creating a graph that when executed will render an entire frame, where nodes represent render passes.
-CopyTextureToTexture command to the command context.
-BitField helper class for Parablit to use with 'enum class' type enums, which should help us eliminate C26812 errors in the project.
-Updated CLib to version 0.2.

Changed:

-All rendering code in main except for the MVP buffer have been moved to a RenderGraphBehaviour called BasicColorPass.

Fixed:

-Swapchain would use a hard-coded imagespace.
-Numerous warnings across the client and ParaBlit projects.
-Allocator would not removed re-used memory blocks from the free list, causing multiple allocations of the same block.
-Render target views would add an invalid descriptor index to the free descriptor index pool when destroyed.

+++ VKR/VKRClient/Mesh.h
@@ -91,8 +91,8 @@ namespace PBClient
 		PB::IRenderer* m_renderer = nullptr;
 		PB::IBufferObject* m_vertexBuffer = nullptr;
 		PB::IBufferObject* m_indexBuffer = nullptr;
-		uint32_t m_totalVertexCount = 0;
-		uint32_t m_totalIndexCount = 0;
+		uint64_t m_totalVertexCount = 0;
+		uint64_t m_totalIndexCount = 0;
 		const char* m_filePath = nullptr;
 		std::string m_name;
 		bool m_empty = true;
